Lab 2 var 10
