+++ 10/lab111/lab1.txt
@@ -1,0 +1,5 @@
+80
+80
+10
+15
+40
